+++ docs/news.md
@@ -4,7 +4,9 @@
 
 **Features:**
 
-- GPU-accelerated kNN searches available.
+- GPU-accelerated kNN searches available -> supporting a GPU-accelerated version
+  of UMAP and tSNE.
+- Added KmKnn nearest neighbour search as a default.
 
 ## 0.1.15